Moved IgniteCacheOffheapManagerImpl's inner classes into separate package.
(cherry picked from commit 4f80f86)
